Update the main display README.md with a more detailed instruction of loading client's own model
I decided to add extra details on how to make the Unity loader work with your model. The text might require additional editing as the author has a bias introduced by an hour and a half of troubleshooting. Maybe it is just the author's problem, as he was thinking in the wrong direction from the beginning of familiarizing himself with the scene. The more technically advanced user of the Unity Loader may be OK.

However, in my personal experience, I hope that the markdown appendix will act as a useful FYI because I was very close to getting discouraged by failing to make this work. It may scare the potential client away. However, I was delighted when I managed to load my model after I understood exactly what I had to do. Such a feeling may hook a potential client by sparking a window of new ideas on getting most of the scene provided.

I apologize for the potential reveal of some internal details that I discovered. Feel free to edit the snippet to your preference, contact me, and request the deletion of my fork.